\name{mrkdat}
\docType{data}
\alias{mrkdat}
\title{A tiny data frame with marker data}
\description{
This data set contains marker data for two individuals 
and two markers in a non-PediHaplotyper format
and is used in the examples fro transpose, 
twocols2tworows and tworows2twocols.
}
\usage{mrkdat}
\format{A data frame with 3 columns and 4 rows.}
% \keyword{datasets}