\name{PediHaplotyper-package}
\alias{PediHaplotyper-package}
\alias{PediHaplotyper}
\docType{package}
\title{
\packageTitle{PediHaplotyper}
}
\description{
\packageDescription{PediHaplotyper}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PediHaplotyper}
\packageIndices{PediHaplotyper}
PediHaplotyper has two functions to perform the haploblock allele assignment.
Function haplotyping_session is used when the input is in generic 
(tab-separated text file) format, while fq_haplotyping_session is used with
input data in FlexQTL format.
Three functions are provided to re-format the phased marker genotype data to
the required format (one row per individual, two columns per marker) if
necessary: transtose, twocols2tworows and tworows2twocols.
}
\author{
\packageAuthor{PediHaplotyper}

Maintainer: \packageMaintainer{PediHaplotyper}
}
\references{
% ~~ Literature or other references for background information ~~
}
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
\keyword{ package }
\seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
% ~~ simple examples of the most important functions ~~
}
